pyproject.toml
[tool.poetry]
name = "remla"
version = "0.1.0"
description = ""
authors = ["Zak Espley <zespley@gmail.com>"]
readme = "README.md"
include = [{path="setup", format=["sdist", "wheel"]}]

[tool.poetry.scripts]
remla = "remla.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.12.1"}
ruamel-yaml = "^0.18.6"
rpistepper = "^0.3a1"
pyvisa = "^1.14.1"
pyvisa-py = "^0.7.2"
adafruit-circuitpython-motorkit = "^1.6.14"
python-tplink-smarthome = "^0.1.3"
dlipower = "^1.0.176"
websockets = "^12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
